requirements.txt
flask
flask-sqlalchemy
flask-cors
werkzeug
deta
requests
pyjwt